requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
dateparser==1.3.0
frozenlist==1.8.0
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.1
propcache==0.4.1
pycryptodome==3.23.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
python_binance==1.0.36
pytz==2026.1.post1
regex==2026.2.28
requests==2.33.0
rich==14.3.3
shellingham==1.5.4
six==1.17.0
typer==0.24.1
tzlocal==5.3.1
urllib3==2.6.3
websockets==16.0
yarl==1.23.0
